totally stucked. now the problem is to solve packet driver problme and the problme with how to imply header sorting problme.
and alos the destnation and address,
puls the sory address problme.

i think the pronlme is the find the the sor or port address first, then  try to find the prost number and match, then open  while lliip is any of the sort number is there, then tyr to get i will tyr the while loop and logicla filtering. then we will see.